test: Add test for deriveFilterDecorations
